[Rename] drop이 sql 예약어인 관계로 waterdrop으로 변경
