+++ src/main/java/com/GDGoC/BaS/user/User.java
@@ -1,28 +1,32 @@
 package com.GDGoC.BaS.user;
 
+import static jakarta.persistence.CascadeType.ALL;
+import static jakarta.persistence.EnumType.STRING;
+import static jakarta.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PROTECTED;
+
 import com.GDGoC.BaS.booth.BoothUser;
 import com.GDGoC.BaS.clothing.UserAccessory;
 import com.GDGoC.BaS.clothing.UserHead;
 import com.GDGoC.BaS.clothing.UserTowel;
-import com.GDGoC.BaS.drop.DropHistory;
+import com.GDGoC.BaS.waterdrop.WaterdropHistory;
 import com.GDGoC.BaS.notification.Notification;
 import com.GDGoC.BaS.shower.UserRecord;
 import com.GDGoC.BaS.user.enums.Eye;
 import com.GDGoC.BaS.user.enums.Mouth;
 import com.GDGoC.BaS.user.enums.Nose;
 import com.GDGoC.BaS.user.enums.Skin;
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
-import lombok.NoArgsConstructor;
-
 import java.util.ArrayList;
 import java.util.List;
-
-import static jakarta.persistence.CascadeType.ALL;
-import static jakarta.persistence.EnumType.STRING;
-import static jakarta.persistence.GenerationType.IDENTITY;
-import static lombok.AccessLevel.PROTECTED;
+import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = PROTECTED)
@@ -45,7 +49,7 @@ public class User {
 
     @Column(nullable = false, columnDefinition = "INT DEFAULT 0")
     @Min(0)
-    private Integer drop;
+    private Integer waterdrop;
 
     @Enumerated(STRING)
     @Column(nullable = false, length = 20, columnDefinition = "VARCHAR(20) DEFAULT 'LIGHT'")
@@ -64,7 +68,7 @@ public class User {
     private Mouth mouth;
 
     @OneToMany(mappedBy = "user", cascade = ALL, orphanRemoval = true)
-    private List<DropHistory> dropHistories = new ArrayList<>();
+    private List<WaterdropHistory> waterdropHistories = new ArrayList<>();
 
     @OneToMany(mappedBy = "user", cascade = ALL, orphanRemoval = true)
     private List<UserRecord> userRecords = new ArrayList<>();
