+++ src/main/java/com/GDGoC/BaS/waterdrop/WaterdropHistory.java
@@ -1,25 +1,32 @@
-package com.GDGoC.BaS.drop;
-
-import com.GDGoC.BaS.user.User;
-import jakarta.persistence.*;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDate;
+package com.GDGoC.BaS.waterdrop;
 
 import static jakarta.persistence.EnumType.STRING;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
+import com.GDGoC.BaS.user.User;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import java.time.LocalDate;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
 @Entity
 @EntityListeners(AuditingEntityListener.class)
 @NoArgsConstructor(access = PROTECTED)
-public class DropHistory {
+public class WaterdropHistory {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    private Long dropHistoryId;
+    private Long waterdropHistoryId;
 
     @CreatedDate
     @Column(nullable = false)
